+++ apps/sim/executor/handlers/pi/cloud/event-filter-source.ts
@@ -1,0 +1,142 @@
+/**
+ * The stdout filter the sandbox modes pipe the Pi CLI through, written at runtime like the search
+ * extension and the review tools script next to it.
+ *
+ * Pi's `--mode json` writes `JSON.stringify(event)` for every session event with no filtering, and
+ * `message_update` repeats the whole assistant message alongside each delta — so raw stdout grows
+ * with the square of the response length, and `tool_execution_end`, `turn_end`, and `agent_end`
+ * each add a full tool result, turn transcript, or run transcript on top of that.
+ *
+ * The reduction has to happen in the sandbox because by the time Sim could drop the bytes they are
+ * already retained: E2B's SDK accumulates every callback-delivered chunk internally, so its adapter
+ * charges all of it against `MAX_SANDBOX_PROCESS_OUTPUT_BYTES` even though the caller streams it,
+ * and crossing that limit kills the sandbox mid-run — which on Create PR loses finished agent work
+ * before the push. Daytona keeps only a rolling tail of a streamed output, so there the same stream
+ * costs bandwidth and parse time rather than the run.
+ */
+
+/**
+ * Written outside `REPO_DIR` so `git add -A` cannot stage it into the user's pull request. The
+ * agent can still read or overwrite it — it holds bash on the sandbox — but the pipeline loads it
+ * at startup, before the agent's first turn, so a rewrite cannot change the events of the
+ * invocation that read it.
+ */
+export const PI_EVENT_FILTER_PATH = '/workspace/sim-pi-event-filter.mjs'
+
+/**
+ * Every emitted shape is a subset of the fields `normalizePiEvent` reads, so the events Sim parses
+ * are identical to the unfiltered stream's — an event it would normalize to `other` is dropped
+ * outright, since nothing downstream distinguishes those from an absent line.
+ *
+ * That makes the two modules a pair: a field added to `normalizePiEvent` has to be added here too,
+ * or it arrives only on the local and review paths while the cloud backends silently lose it.
+ * `shared.test.ts` pins the equivalence by running this filter and comparing normalized output
+ * against the normalized raw events, for the shapes it covers.
+ */
+export const PI_EVENT_FILTER_SOURCE = `/**
+ * Generated by Sim. Compacts the Pi CLI's JSON event stream before the sandbox provider retains it.
+ * Mirrors apps/sim/executor/handlers/pi/cloud/event-filter-source.ts.
+ */
+
+function asRecord(value) {
+  return typeof value === 'object' && value !== null && !Array.isArray(value) ? value : null
+}
+
+function asString(value) {
+  return typeof value === 'string' ? value : undefined
+}
+
+function compactUsage(value) {
+  const usage = asRecord(value)
+  if (!usage) return null
+  return {
+    input: usage.input,
+    output: usage.output,
+    inputTokens: usage.inputTokens,
+    outputTokens: usage.outputTokens,
+    prompt_tokens: usage.prompt_tokens,
+    completion_tokens: usage.completion_tokens,
+  }
+}
+
+function compactAssistantMessage(value) {
+  const message = asRecord(value)
+  if (!message || message.role !== 'assistant') return null
+  return {
+    role: 'assistant',
+    stopReason: asString(message.stopReason),
+    errorMessage: asString(message.errorMessage),
+  }
+}
+
+function compactEvent(value) {
+  const event = asRecord(value)
+  if (!event) return null
+
+  switch (event.type) {
+    case 'message_update': {
+      const update = asRecord(event.assistantMessageEvent)
+      if (update?.type !== 'text_delta' && update?.type !== 'thinking_delta') return null
+      return {
+        type: 'message_update',
+        assistantMessageEvent: { type: update.type, delta: asString(update.delta) },
+      }
+    }
+    case 'tool_execution_start':
+      return { type: 'tool_execution_start', toolName: asString(event.toolName) }
+    case 'tool_execution_end':
+      return {
+        type: 'tool_execution_end',
+        toolName: asString(event.toolName),
+        isError: event.isError === true,
+      }
+    case 'turn_end': {
+      const message = asRecord(event.message)
+      const usage = compactUsage(event.usage)
+      const messageUsage = compactUsage(message?.usage)
+      if (!usage && !messageUsage) return null
+      return {
+        type: 'turn_end',
+        ...(usage ? { usage } : {}),
+        ...(messageUsage ? { message: { usage: messageUsage } } : {}),
+      }
+    }
+    case 'agent_end': {
+      if (event.willRetry === true) return { type: 'agent_end', willRetry: true }
+      const messages = Array.isArray(event.messages) ? event.messages : []
+      let assistant = null
+      for (let index = messages.length - 1; index >= 0; index -= 1) {
+        assistant = compactAssistantMessage(messages[index])
+        if (assistant) break
+      }
+      return { type: 'agent_end', messages: assistant ? [assistant] : [] }
+    }
+    case 'error':
+      return {
+        type: 'error',
+        error: asString(event.error),
+        message: asString(event.message),
+      }
+    default:
+      return null
+  }
+}
+
+function processLine(line) {
+  if (!line.trim()) return
+  try {
+    const event = compactEvent(JSON.parse(line))
+    if (event) process.stdout.write(JSON.stringify(event) + '\\n')
+  } catch {}
+}
+
+process.stdin.setEncoding('utf8')
+let buffer = ''
+process.stdin.on('data', (chunk) => {
+  buffer += chunk
+  const lines = buffer.split('\\n')
+  buffer = lines.pop() ?? ''
+  for (const line of lines) processLine(line)
+})
+process.stdin.on('end', () => processLine(buffer))
+`
